Added generic method to read a tree object using a parent/child navigation.
