+++ MetasysServices/MetasysClient.cs
@@ -16,11 +16,8 @@ namespace JohnsonControls.Metasys.BasicServices
     /// <summary>
     /// An HTTP client for consuming the most commonly used endpoints of the Metasys API.
     /// </summary>
-    public class MetasysClient : IMetasysClient
-    {
-        /// <summary>The http client.</summary>
-        protected FlurlClient Client;
-
+    public class MetasysClient : BasicServiceProvider, IMetasysClient
+    {   
         /// <summary>The current session token.</summary>
         protected AccessToken AccessToken;
 
@@ -48,6 +45,11 @@ namespace JohnsonControls.Metasys.BasicServices
         /// Stores retrieved Ids and serves as an in-memory caching layer
         /// </summary>
         protected Dictionary<string, Guid> IdentifiersDictionary = new Dictionary<string, Guid>();
+
+        /// <summary>
+        /// Local instance of Trends service
+        /// </summary>
+        public ITrendsService Trends;
 
         /// <summary>
         /// Creates a new MetasysClient.
@@ -85,6 +87,8 @@ namespace JohnsonControls.Metasys.BasicServices
                 Client = new FlurlClient($"https://{hostname}"
                     .AppendPathSegments("api", version));
             }
+            // Init related services
+            Trends = new TrendsServiceProvider(Client);            
         }
 
         /// <summary>
@@ -267,31 +271,7 @@ namespace JohnsonControls.Metasys.BasicServices
             ObjectTypeEnumerations.Add(ObjectTypeEnumerations2);
         }
 
-        /// <summary>
-        /// Throws a Metasys Exception from a Flurl.Http exception.
-        /// </summary>
-        /// <exception cref="MetasysHttpParsingException"></exception>
-        /// <exception cref="MetasysHttpTimeoutException"></exception>
-        /// <exception cref="MetasysHttpException"></exception>
-        protected void ThrowHttpException(Flurl.Http.FlurlHttpException e)
-        {
-            if (e.Call.Response.StatusCode == HttpStatusCode.NotFound)
-            {
-                throw new MetasysHttpNotFoundException(e);
-            }
-            if (e.GetType() == typeof(Flurl.Http.FlurlParsingException))
-            {
-                throw new MetasysHttpParsingException((Flurl.Http.FlurlParsingException)e);
-            }
-            else if (e.GetType() == typeof(Flurl.Http.FlurlHttpTimeoutException))
-            {
-                throw new MetasysHttpTimeoutException((Flurl.Http.FlurlHttpTimeoutException)e);
-            }
-            else
-            {
-                throw new MetasysHttpException(e);
-            }
-        }
+    
 
         /// <summary>Use to log an error message from an asynchronous context.</summary>
         private async Task LogErrorAsync(String message)
@@ -495,29 +475,7 @@ namespace JohnsonControls.Metasys.BasicServices
                 }
             }
             return IdentifiersDictionary[normalizedItemReference];
-        }
-
-        /// <summary>
-        /// Parses a JToken and creates a Guid.
-        /// </summary>
-        /// <param name="token"></param>
-        /// <exception cref="MetasysGuidException"></exception>
-        /// <returns>A Guid representation of the JToken.</returns>
-        protected Guid ParseObjectIdentifier(JToken token)
-        {
-            string str = null;
-            try
-            {
-                str = token.Value<string>();
-                var id = new Guid(str);
-                return id;
-            }
-            catch (Exception e) when (e is System.ArgumentNullException ||
-                e is System.ArgumentException || e is System.FormatException)
-            {
-                throw new MetasysGuidException(str, e);
-            }
-        }
+        }       
 
         /// <summary>
         /// Read one attribute value given the Guid of the object.
@@ -906,40 +864,7 @@ namespace JohnsonControls.Metasys.BasicServices
         public async Task<IEnumerable<MetasysObject>> GetNetworkDevicesAsync(string type = null)
         {
             return await ProcessPagedRequestAsync("networkDevices", type);
-        }
-
-        /// <summary>
-        /// Generic paged request for the given resource and optional type
-        /// </summary>
-        /// <param name="resource"></param>
-        /// <param name="type"></param>
-        /// <param name="page"></param>
-        /// <exception cref="MetasysHttpException"></exception>
-        protected async Task<JToken> PagedRequestAsync(string resource, string type = null, int page = 1, int pageSize=100)
-        {
-            Url url = new Url(resource);
-            url.SetQueryParam("page", page);
-            url.SetQueryParam("pageSize", page);
-            if (type != null)
-            {
-                url.SetQueryParam("type", type);
-            }
-
-            try
-            {
-                var response = await Client.Request(url)
-                    .GetJsonAsync<JToken>()
-                    .ConfigureAwait(false);
-
-                return response;
-            }
-            catch (FlurlHttpException e)
-            {
-                ThrowHttpException(e);
-            }
-
-            return null;
-        }
+        }      
 
 
         /// <summary>
@@ -1076,129 +1001,11 @@ namespace JohnsonControls.Metasys.BasicServices
         /// <exception cref="MetasysHttpException"></exception>
         /// <exception cref="MetasysHttpParsingException"></exception>
 
-        public IEnumerable<MetasysObject> GetObjects(Guid id, int levels = 1, string parentResource = "objects", string childResource = "objects")
+        public IEnumerable<MetasysObject> GetObjects(Guid id, int levels = 1)
         {
             return GetObjectsAsync(id, levels).GetAwaiter().GetResult();
-        }
-
-        /// <summary>
-        /// Gets all child objects given a parent Guid asynchronously by requesting each available page.
-        /// Level indicates how deep to retrieve objects.
-        /// </summary>
-        /// <remarks>
-        /// A level of 1 only retrieves immediate children of the parent object.
-        /// </remarks>
-        /// <param name="id"></param>
-        /// <param name="levels">The depth of the children to retrieve.</param>    
-        /// <param name="parentResource">The parent resource to retrieve children.</param>    
-        /// <param name="childResource">The children resource to get related elements</param>    
-        /// <exception cref="MetasysHttpException"></exception>
-        /// <exception cref="MetasysHttpParsingException"></exception>
-        public async Task<IEnumerable<MetasysObject>> GetObjectsAsync(Guid id, int levels = 1, string parentResource="objects", string childResource = "objects")
-        {
-            if (levels < 1)
-            {
-                return null;
-            }
-
-            List<MetasysObject> objects = new List<MetasysObject>() { };
-            bool hasNext = true;
-            int page = 1;
-
-            while (hasNext)
-            {
-                hasNext = false;
-                var response = await GetObjectsRequestAsync(id, page, parentResource, childResource).ConfigureAwait(false);
-                if (response == null || response.Type == JTokenType.Null || !response.HasValues)
-
-                {
-
-                    return null;
-
-                }
-                try
-                {
-                    var total = response["total"].Value<int>();
-                    if (total > 0)
-                    {
-                        var list = response["items"] as JArray;
-
-                        foreach (var item in list)
-                        {
-                            try
-                            {
-                                IEnumerable<MetasysObject> children = null;
-                                if (levels - 1 > 0)
-                                {
-                                    var objId = ParseObjectIdentifier(item["id"]);
-                                    if (objId != null)
-                                    {
-                                        try
-                                        {
-                                            children = await GetObjectsAsync(objId, levels - 1).ConfigureAwait(false);
-                                        }
-                                        catch (Exception e) when (e is MetasysObjectException ||
-                                            e is MetasysHttpParsingException)
-                                        {
-                                            throw new MetasysObjectException(e);
-                                        }
-                                    }
-                                }
-                                MetasysObject obj = new MetasysObject(item, children);
-                                objects.Add(obj);
-                            }
-                            catch (MetasysGuidException)
-                            {
-                                // Error with this item's id, skip and do not create object
-                            }
-                            catch (MetasysObjectException e)
-                            {
-                                // There was an issue creating the object
-                                throw e;
-                            }
-                        }
-
-                        if (response["next"] != null && response["next"].Type != JTokenType.Null)
-                        {
-                            hasNext = true;
-                            page++;
-                        }
-                    }
-                }
-                catch (System.NullReferenceException e)
-                {
-                    throw new MetasysHttpParsingException(response.ToString(), e);
-                }
-            }
-            return objects;
-        }
-
-        /// <summary>
-        /// Gets all child objects given a parent Guid asynchronously with the given page number.
-        /// </summary>
-        /// <param name="id"></param>
-        /// <param name="page"></param>
-        /// <param name="parentResource">The parent resource to retrieve children.</param>    
-        /// <param name="childResource">The children resource to get related elements.</param>    
-        /// <exception cref="MetasysHttpException"></exception>
-        private async Task<JToken> GetObjectsRequestAsync(Guid id, int page = 1, string parentResource = "objects", string childResource="objects" )
-        {
-            Url url = new Url(parentResource)
-                .AppendPathSegments(id, childResource)
-                .SetQueryParam("page", page);
-            JToken response = null;
-            try
-            {
-                response = await Client.Request(url)
-                .GetJsonAsync<JToken>()
-                .ConfigureAwait(false);                          
-            }
-            catch (FlurlHttpException e)
-            {
-                ThrowHttpException(e);
-            }
-            return response;
-        }
+        }       
+       
 
         /// <summary>
         /// Gets all spaces by requesting each available page.
@@ -1222,61 +1029,7 @@ namespace JohnsonControls.Metasys.BasicServices
             return await ProcessPagedRequestAsync("spaces", type);
         }
 
-        /// <summary>
-        /// Gets all items for the given resource asynchronously by requesting each available page.
-        /// </summary>
-        /// <param name="resource"></param>
-        /// <param name="type">Optional type number as a string</param>
-        /// <param name="skip">Optional how many page we skip. 0 by default.</param>
-        /// <param name="take">Optional how many page we take. 0 for all.</param>
-        /// <exception cref="MetasysHttpException"></exception>
-        /// <exception cref="MetasysHttpParsingException"></exception>
-        protected async Task<IEnumerable<MetasysObject>> ProcessPagedRequestAsync(string resource, string type = null, PageSettings settings = null)
-        {
-            List<MetasysObject> items = new List<MetasysObject>() { };
-            bool hasNext = true;
-            settings = settings ?? new PageSettings { Skip=0, Take=0, PageSize=100}; // Set default values for paging
-            int page = settings.Skip +1;
-            int i = 0;
-            while (hasNext && (settings.Take==0 || i<settings.Take))
-            {
-                hasNext = false;
-                var response = await PagedRequestAsync(resource, type, page,settings.PageSize).ConfigureAwait(false);
-                try
-                {
-                    var total = response["total"].Value<int>();
-                    if (total > 0)
-                    {
-                        var list = response["items"] as JArray;
-                        foreach (var item in list)
-                        {
-                            try
-                            {
-                                MetasysObject space = new MetasysObject(item);
-                                items.Add(space);
-                            }
-                            catch (MetasysObjectException e)
-                            {
-                                throw e;
-                            }
-                        }
-
-                        if (!(response["next"] == null || response["next"].Type == JTokenType.Null))
-                        {
-                            hasNext = true;
-                            page++;
-                            i++;
-                        }
-                    }
-                }
-                catch (System.NullReferenceException e)
-                {
-                    throw new MetasysHttpParsingException(response.ToString(), e);
-                }
-            }
-            return items;
-        }
-
+      
         /// <summary>
         /// Gets all equipment by requesting each available page.
         /// </summary>
@@ -1330,7 +1083,7 @@ namespace JohnsonControls.Metasys.BasicServices
         /// <returns></returns>
         public async Task<IEnumerable<MetasysObject>> GetSpaceEquipmentAsync(Guid spaceId)
         {
-            return await GetObjectsAsync(spaceId, parentResource: "spaces", childResource: "equipment");
+            return toMetasysObject(await GetObjectsAsync(spaceId, parentResource: "spaces", childResource: "equipment"));
         }
 
         /// <summary>
@@ -1399,6 +1152,11 @@ namespace JohnsonControls.Metasys.BasicServices
                 pointsWithValues.Add(point);
             }       
             return pointsWithValues;          
+        }
+
+        public async Task<IEnumerable<MetasysObject>> GetObjectsAsync(Guid id, int levels)
+        {
+            return toMetasysObject(await GetObjectsAsync(id, "objects","objects",levels));
         }
     }
 }
