+++ MetasysServices/Trends/Interfaces/ITrendsService.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace JohnsonControls.Metasys.BasicServices
+{
+    public interface ITrendsService
+    {
+        List<Guid>GetTrendedAttributes(Guid id);
+        List<Sample> GetSamples(Guid objectId, Guid attributeId, DateTime startTime, DateTime endTime);
+    }
+}
